refactor: 가격 조정 Slider 컴포넌트 분리
