+++ src/App.tsx
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { RouterProvider } from 'react-router-dom';
 
+import router from './Router';
 import { store } from './redux/store';
-import router from './utils/Router';
 
 const theme = extendTheme({
   colors: {
