+++ src/components/filters/Price.tsx
@@ -1,5 +1,4 @@
 import {
-  Badge,
   Button,
   Flex,
   Modal,
@@ -8,25 +7,13 @@ import {
   ModalContent,
   ModalHeader,
   ModalOverlay,
-  RangeSlider,
-  RangeSliderFilledTrack,
-  RangeSliderThumb,
-  RangeSliderTrack,
   Text,
   useDisclosure,
 } from '@chakra-ui/react';
 
-import { useAppDispatch, useAppSelector } from '../../redux/hook/redux.hook';
-import { maxPrice, priceStep, priceSteps } from '../../redux/redux.interface';
-import { price } from '../../redux/slice/productslice';
+import PriceRangeSlider from './PriceRangeSlider';
 
 const PriceFilter = () => {
-  const priceFilter = useAppSelector((state) => state.product.priceFilter);
-
-  const dispatch = useAppDispatch();
-  const sliderChange = (priceRange: number[]) => {
-    dispatch(price(priceRange));
-  };
   const { isOpen, onOpen, onClose } = useDisclosure();
 
   return (
@@ -44,34 +31,8 @@ const PriceFilter = () => {
           <ModalBody>
             <Flex mx={10} mb={5} flexDir={'column'}>
               <Text mb={5}>선택하신 가격 범위</Text>
-              <RangeSlider
-                w="full"
-                aria-label={['min', 'max']}
-                colorScheme="facebook" // 색상 바꾸기
-                defaultValue={[priceFilter.min, priceFilter.max]}
-                min={0}
-                max={maxPrice}
-                step={priceStep}
-                minStepsBetweenThumbs={10}
-                onChangeEnd={(e) => sliderChange(e)}
-              >
-                <RangeSliderTrack>
-                  <RangeSliderFilledTrack />
-                </RangeSliderTrack>
-
-                <RangeSliderThumb index={0}></RangeSliderThumb>
-                <RangeSliderThumb index={1}></RangeSliderThumb>
-              </RangeSlider>
-              <Flex justifyContent={'space-between'}>
-                {priceSteps.map((step, index) => {
-                  return (
-                    <Badge key={index} mt="5px" fontSize="sm" color="black">
-                      ₩{step}
-                    </Badge>
-                  );
-                })}
-              </Flex>
-            </Flex>{' '}
+              <PriceRangeSlider />
+            </Flex>
           </ModalBody>
         </ModalContent>
       </Modal>
